+++ src/main/java/insane96mcp/iguanatweaksreborn/setup/client/ClientSetup.java
@@ -94,7 +94,7 @@ public class ClientSetup {
                 addAfter(event, Items.NETHER_GOLD_ORE, CoalFire.SOUL_SOIL_HELLISH_COAL_ORE.item().get());
                 addAfter(event, Items.NETHER_GOLD_ORE, CoalFire.SOUL_SAND_HELLISH_COAL_ORE.item().get());
             }
-            if (Feature.isEnabled(FlintExpansion.class))
+            if (Feature.isEnabled(FlintExpansion.class) && !FlintExpansion.dontEnableFlintTools)
                 addAfter(event, Items.CACTUS, FlintExpansion.FLINT_ROCK.item().get());
         }
         else if (event.getTabKey() == CreativeModeTabs.TOOLS_AND_UTILITIES) {
@@ -104,7 +104,7 @@ public class ClientSetup {
                 addAfter(event, Items.WOODEN_HOE, CopperEquipment.PICKAXE.get());
                 addAfter(event, Items.WOODEN_HOE, CopperEquipment.SHOVEL.get());
             }
-            if (Feature.isEnabled(FlintExpansion.class)) {
+            if (Feature.isEnabled(FlintExpansion.class) && !FlintExpansion.dontEnableFlintTools) {
                 addAfter(event, Items.WOODEN_HOE, FlintExpansion.HOE.get());
                 addAfter(event, Items.WOODEN_HOE, FlintExpansion.AXE.get());
                 addAfter(event, Items.WOODEN_HOE, FlintExpansion.PICKAXE.get());
@@ -134,7 +134,7 @@ public class ClientSetup {
                 addAfter(event, Items.LEATHER_BOOTS, CopperEquipment.CHESTPLATE.get());
                 addAfter(event, Items.LEATHER_BOOTS, CopperEquipment.HELMET.get());
             }
-            if (Feature.isEnabled(FlintExpansion.class)) {
+            if (Feature.isEnabled(FlintExpansion.class) && !FlintExpansion.dontEnableFlintTools) {
                 addAfter(event, Items.WOODEN_SWORD, FlintExpansion.SWORD.get());
                 addAfter(event, Items.WOODEN_AXE, FlintExpansion.AXE.get());